Fix compilation failure by updating Gradle and modernizing API usage.
This change resolves a major build failure caused by an outdated Gradle configuration and code that was incompatible with Minecraft 1.21.8 APIs. It also corrects a logic error where hardcoded values were used in checks instead of real-time data.

The following key changes were made:

- **Gradle Configuration:**
  - Added a `settings.gradle` file to declare the necessary Maven repositories for Fabric and Gradle plugins.
  - Added the Gradle Wrapper (`gradlew`) configured to use Gradle 8.14, ensuring a consistent and compatible build environment.
  - Updated `build.gradle` to use the modern `runConfigs` block instead of the deprecated `splitEnvironmentSourceSets`, which resolved the client-side dependency issues.

- **API and Code Modernization:**
  - Replaced direct field access for previous positions (`prevX`, etc.) with the new `lastX`, `lastY`, and `lastZ` fields in `SimulatedPlayer.java`.
  - Updated the movement speed attribute to `EntityAttributes.MOVEMENT_SPEED`.
  - Implemented a packet-based tracking system for the damage tilt animation (`DamageTiltS2CPacket`) to replace the removed `getDamageTiltAngle()` method. This involved creating a `DamageTiltTracker` and a new mixin to intercept the packet.
  - Corrected various other method calls (`isInsideWaterOrBubbleColumn` to `isSubmergedInWater`) and type mismatches (`float` to `double` for `fallDistance`).

- **Logic Correction:**
  - Integrated the `TpsTracker` into `SpeedCheck.java` and `StrafeCheck.java` to use real-time TPS and ping data, removing the incorrect hardcoded placeholder values.

+++ src/main/java/net/blosson/lflagger/checks/NoFallCheck.java
@@ -11,7 +11,7 @@ public class NoFallCheck extends Check {
     // Inner class to store the state of a player between ticks
     private static class PlayerFallState {
         boolean wasOnGround;
-        float lastFallDistance;
+        double lastFallDistance; // Changed to double
 
         PlayerFallState(PlayerEntity player) {
             this.wasOnGround = player.isOnGround();
@@ -41,7 +41,7 @@ public class NoFallCheck extends Check {
         // Check if the player just landed
         if (!state.wasOnGround && isCurrentlyOnGround) {
             // A fall distance of > 3.0 is required for damage. We use 4.0 for a conservative check.
-            if (state.lastFallDistance > 4.0f) {
+            if (state.lastFallDistance > 4.0) { // Changed to double comparison
                 // The player fell far enough to take damage.
                 // A player's hurtTime becomes > 0 when they take damage.
                 // If it's 0, they likely negated the damage.
